Added: ColorManager, documentation for the permissions in the plugin.yml
